# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patoughi Driver API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/auth/submit-information": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Driver Submit Information",
                "operationId": "submitInformation",
                "requestBody": {
                    "description": "Driver Update data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDriverRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DriverResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/drivers/check-verify-biometric/{trackId}": {
            "get": {
                "tags": [
                    "Driver"
                ],
                "summary": "Driver Check Verify Biometric",
                "operationId": "driverCheckVerifyBiometric",
                "parameters": [
                    {
                        "name": "trackId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "your biometric validation was successful"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/stations/{station}/loading-locations": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "List all loading locations for a station",
                "operationId": "LoadingLocation",
                "parameters": [
                    {
                        "name": "station",
                        "in": "path",
                        "description": "Station ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LoadingLocationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during loadingLocations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/stations": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "List all stations",
                "operationId": "Station",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/StationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during stations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/{ticket}/reply": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Reply ticket",
                "operationId": "replyTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "The id of the ticket",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Ticket Message Information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TicketReplyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket Reply Submitted Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TicketMessageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateDriverRequest": {
                "required": [
                    "licence_number",
                    "licence_type",
                    "licence_expired_at"
                ],
                "properties": {
                    "licence_number": {
                        "type": "string"
                    },
                    "licence_type": {
                        "type": "string"
                    },
                    "licence_expired_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "emergency_contact_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "emergency_contact_phone": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "TicketReplyRequest": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            },
...
            "DriverResource": {
                "required": [
                    "id",
                    "mdm_organization_id",
                    "first_name",
                    "last_name",
                    "national_code",
                    "gender",
                    "birth_date",
                    "licence_number",
                    "licence_type",
                    "licence_expired_at",
                    "hired_at",
                    "emergency_contact_name",
                    "emergency_contact_phone"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "mdm_organization_id": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "national_code": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "licence_number": {
                        "type": "string"
                    },
                    "licence_type": {
                        "type": "string"
                    },
                    "licence_expired_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "hired_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "emergency_contact_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "emergency_contact_phone": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "LoadingLocationResource": {
                "required": [
                    "id",
                    "mdm_organization_id",
                    "status",
                    "name",
                    "description",
                    "station_id",
                    "usage",
                    "driver_instruction"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "mdm_organization_id": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "station_id": {
                        "type": "string"
                    },
                    "usage": {
                        "type": "string"
                    },
                    "driver_instruction": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "StationResource": {
                "required": [
                    "id",
                    "mdm_organization_id",
                    "status",
                    "name",
                    "description",
                    "carrier_id",
                    "short_name",
                    "public_name",
                    "start_opening_time",
                    "end_opening_hour",
                    "injection_days",
                    "required_time_slot",
                    "address_country",
                    "address_city",
                    "address_state",
                    "address_street_and_house_number",
                    "address_postal_code",
                    "address_lat",
                    "address_lng",
                    "availability",
                    "stop_type",
                    "vehicle_restriction"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "mdm_organization_id": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "carrier_id": {
                        "type": "string"
                    },
                    "short_name": {
                        "type": "string"
                    },
                    "public_name": {
                        "type": "string"
                    },
                    "start_opening_time": {
                        "type": "time"
                    },
                    "end_opening_hour": {
                        "type": "time"
                    },
                    "injection_days": {
                        "type": "string"
                    },
                    "required_time_slot": {
                        "type": "boolean"
                    },
                    "address_country": {
                        "type": "string"
                    },
                    "address_city": {
                        "type": "string"
                    },
                    "address_state": {
                        "type": "string"
                    },
                    "address_street_and_house_number": {
                        "type": "string"
                    },
                    "address_postal_code": {
                        "type": "string"
                    },
                    "address_lat": {
                        "type": "decimal"
                    },
                    "address_lng": {
                        "type": "decimal"
                    },
                    "availability": {
                        "type": "string"
                    },
                    "stop_type": {
                        "type": "string"
                    },
                    "vehicle_restriction": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TicketMessageResource": {
                "required": [
                    "id",
                    "message",
                    "sender_id",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "sender_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ApiResponseErrorValidation": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ApiResponseErrorServer": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}